+++ appcheck/firebase-appcheck/src/main/java/com/google/firebase/appcheck/internal/DefaultTokenRefresher.java
@@ -18,10 +18,11 @@ import static com.google.android.gms.common.internal.Preconditions.checkNotNull;
 import static java.util.concurrent.TimeUnit.MILLISECONDS;
 import static java.util.concurrent.TimeUnit.SECONDS;
 
-import android.annotation.SuppressLint;
 import androidx.annotation.NonNull;
 import androidx.annotation.VisibleForTesting;
-import java.util.concurrent.Executors;
+import com.google.firebase.annotations.concurrent.Blocking;
+import com.google.firebase.annotations.concurrent.Lightweight;
+import java.util.concurrent.Executor;
 import java.util.concurrent.ScheduledExecutorService;
 import java.util.concurrent.ScheduledFuture;
 
@@ -36,21 +37,18 @@ public class DefaultTokenRefresher {
   @VisibleForTesting static final long MAX_DELAY_SECONDS = 16 * 60; // 16 minutes
 
   private final DefaultFirebaseAppCheck firebaseAppCheck;
+  private final Executor liteExecutor;
   private final ScheduledExecutorService scheduledExecutorService;
 
   private volatile ScheduledFuture<?> refreshFuture;
   private volatile long delayAfterFailureSeconds;
 
-  // TODO(b/258273630): Migrate to go/firebase-android-executors
-  @SuppressLint("ThreadPoolCreation")
-  DefaultTokenRefresher(@NonNull DefaultFirebaseAppCheck firebaseAppCheck) {
-    this(checkNotNull(firebaseAppCheck), Executors.newScheduledThreadPool(/* corePoolSize= */ 1));
-  }
-
-  @VisibleForTesting
   DefaultTokenRefresher(
-      DefaultFirebaseAppCheck firebaseAppCheck, ScheduledExecutorService scheduledExecutorService) {
-    this.firebaseAppCheck = firebaseAppCheck;
+      @NonNull DefaultFirebaseAppCheck firebaseAppCheck,
+      @Lightweight Executor liteExecutor,
+      @Blocking ScheduledExecutorService scheduledExecutorService) {
+    this.firebaseAppCheck = checkNotNull(firebaseAppCheck);
+    this.liteExecutor = liteExecutor;
     this.scheduledExecutorService = scheduledExecutorService;
     this.delayAfterFailureSeconds = UNSET_DELAY;
   }
@@ -92,12 +90,10 @@ public class DefaultTokenRefresher {
     }
   }
 
-  // TODO(b/261013814): Use an explicit executor in continuations.
-  @SuppressLint("TaskMainThread")
   private void onRefresh() {
     firebaseAppCheck
         .fetchTokenFromProvider()
-        .addOnFailureListener(e -> scheduleRefreshAfterFailure());
+        .addOnFailureListener(liteExecutor, e -> scheduleRefreshAfterFailure());
   }
 
   /** Cancels the in-flight scheduled refresh. */