+++ buildSrc/src/main/java/com/google/firebase/gradle/plugins/ci/ChangedModulesTask.kt
@@ -14,6 +14,7 @@
 
 package com.google.firebase.gradle.plugins.ci
 
+import com.google.firebase.gradle.plugins.FirebaseLibraryExtension
 import com.google.gson.Gson
 import java.io.File
 import org.gradle.api.DefaultTask
@@ -22,15 +23,20 @@ import org.gradle.api.tasks.Input
 import org.gradle.api.tasks.OutputFile
 import org.gradle.api.tasks.TaskAction
 import org.gradle.api.tasks.options.Option
+import org.gradle.kotlin.dsl.findByType
 
 abstract class ChangedModulesTask : DefaultTask() {
   @get:Input
-  @set:Option(option = "changed-git-paths", description = "Hellos")
+  @set:Option(option = "changed-git-paths", description = "The list of changed paths")
   abstract var changedGitPaths: List<String>
 
   @get:Input
-  @set:Option(option = "output-file-path", description = "Hello")
+  @set:Option(option = "output-file-path", description = "The file to output json to")
   abstract var outputFilePath: String
+
+  @get:Input
+  @set:Option(option = "only-firebase-sdks", description = "Only list Firebase SDKs")
+  abstract var onlyFirebaseSDKs: Boolean
 
   @get:OutputFile val outputFile by lazy { File(outputFilePath) }
 
@@ -43,6 +49,10 @@ abstract class ChangedModulesTask : DefaultTask() {
     val projects =
       AffectedProjectFinder(project, changedGitPaths.toSet(), listOf())
         .find()
+        .filter {
+          val ext = it.extensions.findByType(FirebaseLibraryExtension::class.java)
+          !onlyFirebaseSDKs || it.extensions.findByType<FirebaseLibraryExtension>() != null
+        }
         .map { it.path }
         .toSet()
 
@@ -50,7 +60,14 @@ abstract class ChangedModulesTask : DefaultTask() {
     project.rootProject.subprojects.forEach { p ->
       p.configurations.forEach { c ->
         c.dependencies.filterIsInstance<ProjectDependency>().forEach {
-          result[it.dependencyProject.path]?.add(p.path)
+          if (
+            !onlyFirebaseSDKs ||
+              it.dependencyProject.extensions.findByType<FirebaseLibraryExtension>() != null
+          ) {
+            if (!onlyFirebaseSDKs || p.extensions.findByType<FirebaseLibraryExtension>() != null) {
+              result[it.dependencyProject.path]?.add(p.path)
+            }
+          }
         }
       }
     }