+++ ci/fireci/fireciplugins/macrobenchmark/run/runner.py
@@ -22,18 +22,25 @@ import yaml
 from .test_project_builder import TestProjectBuilder
 from .utils import execute
 from pathlib import Path
-from typing import Dict
+from typing import Dict, List, Set
 
 
 logger = logging.getLogger('fireci.macrobenchmark')
 
 
-async def start(build_only: bool, local: bool, repeat: int, output: Path):
+async def start(
+    build_only: bool,
+    local: bool,
+    repeat: int,
+    output: Path,
+    changed_modules_file: Path = None
+):
   logger.info('Starting macrobenchmark test ...')
 
   config = _process_config_yaml()
   product_versions = _assemble_all_products()
   test_dir = _prepare_test_directory()
+  changed_traces = _process_changed_modules(changed_modules_file)
   template_project_dir = Path('health-metrics/benchmark/template')
 
   test_projects = [
@@ -42,6 +49,7 @@ async def start(build_only: bool, local: bool, repeat: int, output: Path):
       test_dir,
       template_project_dir,
       product_versions,
+      changed_traces,
     ).build() for test_config in config['test-apps']]
 
   if not build_only:
@@ -99,3 +107,42 @@ def _prepare_test_directory() -> Path:
   test_dir = tempfile.mkdtemp(prefix='benchmark-test-')
   logger.info(f'Temporary test directory created at: {test_dir}')
   return Path(test_dir)
+
+
+def _process_changed_modules(path: Path) -> List[str]:
+  trace_names = {
+    ":appcheck": ["fire-app-check"],
+    ":firebase-abt": ["fire-abt"],
+    ":firebase-appdistribution": ["fire-appdistribution"],
+    ":firebase-config": ["fire-rc"],
+    ":firebase-common": ["Firebase", "ComponentDiscovery", "Runtime"],
+    ":firebase-components": ["Firebase", "ComponentDiscovery", "Runtime"],
+    ":firebase-database": ["fire-rtdb"],
+    ":firebase-datatransport": ["fire-transport"],
+    ":firebase-dynamic-links": ["fire-dl"],
+    ":firebase-crashlytics": ["fire-cls"],
+    ":firebase-crashlytics-ndk": ["fire-cls"],
+    ":firebase-firestore": ["fire-fst"],
+    ":firebase-functions": ["fire-fn"],
+    ":firebase-inappmessaging": ["fire-fiam"],
+    ":firebase-inappmessaging-display": ["fire-fiamd"],
+    ":firebase-installations": ["fire-installations"],
+    ":firebase-installations-interop": ["fire-installations"],
+    ":firebase-messaging": ["fire-fcm"],
+    ":firebase-messaging-directboot": ["fire-fcm"],
+    ":firebase-ml-modeldownloader": ["firebase-ml-modeldownloader"],
+    ":firebase-perf": ["fire-perf"],
+    ":firebase-storage": ["fire-gcs"],
+    ":transport": ["fire-transport"],
+  }
+
+  results: Set[str] = set()
+  if path:
+    with open(path) as changed_modules_file:
+      changed_modules = json.load(changed_modules_file)
+      for module in changed_modules:
+        for product in trace_names:
+          if module.startswith(product):
+            results.update(trace_names[product])
+  logger.info(f"Extracted changed traces {results} from {path}")
+  return list(results)