+++ firebase-ml-modeldownloader/src/main/java/com/google/firebase/ml/modeldownloader/FirebaseModelDownloader.java
@@ -13,7 +13,6 @@
 // limitations under the License.
 package com.google.firebase.ml.modeldownloader;
 
-import android.annotation.SuppressLint;
 import android.os.Build.VERSION_CODES;
 import android.util.Log;
 import androidx.annotation.NonNull;
@@ -26,6 +25,8 @@ import com.google.android.gms.tasks.TaskCompletionSource;
 import com.google.android.gms.tasks.Tasks;
 import com.google.firebase.FirebaseApp;
 import com.google.firebase.FirebaseOptions;
+import com.google.firebase.annotations.concurrent.Background;
+import com.google.firebase.annotations.concurrent.Blocking;
 import com.google.firebase.ml.modeldownloader.internal.CustomModelDownloadService;
 import com.google.firebase.ml.modeldownloader.internal.FirebaseMlLogEvent.ModelDownloadLogEvent.DownloadStatus;
 import com.google.firebase.ml.modeldownloader.internal.FirebaseMlLogEvent.ModelDownloadLogEvent.ErrorCode;
@@ -36,7 +37,6 @@ import com.google.firebase.ml.modeldownloader.internal.SharedPreferencesUtil;
 import java.io.File;
 import java.util.Set;
 import java.util.concurrent.Executor;
-import java.util.concurrent.Executors;
 import javax.inject.Inject;
 
 public class FirebaseModelDownloader {
@@ -47,35 +47,15 @@ public class FirebaseModelDownloader {
   private final ModelFileDownloadService fileDownloadService;
   private final ModelFileManager fileManager;
   private final CustomModelDownloadService modelDownloadService;
-  private final Executor executor;
+  private final Executor bgExecutor;
+  private final Executor blockingExecutor;
 
   private final FirebaseMlLogger eventLogger;
   private final CustomModel.Factory modelFactory;
 
   @Inject
-  @RequiresApi(api = VERSION_CODES.KITKAT)
-  // TODO(b/258424267): Migrate to go/firebase-android-executors
-  @SuppressLint("ThreadPoolCreation")
-  FirebaseModelDownloader(
-      FirebaseOptions firebaseOptions,
-      SharedPreferencesUtil sharedPreferencesUtil,
-      ModelFileDownloadService fileDownloadService,
-      CustomModelDownloadService modelDownloadService,
-      ModelFileManager fileManager,
-      FirebaseMlLogger eventLogger,
-      CustomModel.Factory modelFactory) {
-    this(
-        firebaseOptions,
-        sharedPreferencesUtil,
-        fileDownloadService,
-        modelDownloadService,
-        fileManager,
-        eventLogger,
-        Executors.newSingleThreadExecutor(),
-        modelFactory);
-  }
-
   @VisibleForTesting
+  @RequiresApi(api = VERSION_CODES.KITKAT)
   FirebaseModelDownloader(
       FirebaseOptions firebaseOptions,
       SharedPreferencesUtil sharedPreferencesUtil,
@@ -83,7 +63,8 @@ public class FirebaseModelDownloader {
       CustomModelDownloadService modelDownloadService,
       ModelFileManager fileManager,
       FirebaseMlLogger eventLogger,
-      Executor executor,
+      @Background Executor bgExecutor,
+      @Blocking Executor blockingExecutor,
       CustomModel.Factory modelFactory) {
     this.firebaseOptions = firebaseOptions;
     this.sharedPreferencesUtil = sharedPreferencesUtil;
@@ -91,7 +72,8 @@ public class FirebaseModelDownloader {
     this.modelDownloadService = modelDownloadService;
     this.fileManager = fileManager;
     this.eventLogger = eventLogger;
-    this.executor = executor;
+    this.bgExecutor = bgExecutor;
+    this.blockingExecutor = blockingExecutor;
     this.modelFactory = modelFactory;
   }
 
@@ -227,7 +209,7 @@ public class FirebaseModelDownloader {
 
       if (downloadInProgressTask != null) {
         return downloadInProgressTask.continueWithTask(
-            executor,
+            bgExecutor,
             downloadTask -> {
               if (downloadTask.isSuccessful()) {
                 return finishModelDownload(model.getName());
@@ -251,7 +233,7 @@ public class FirebaseModelDownloader {
     // bad model state - delete all existing model details and return exception
     return deleteDownloadedModel(model.getName())
         .continueWithTask(
-            executor,
+            bgExecutor,
             deletionTask ->
                 Tasks.forException(
                     new FirebaseMlException(
@@ -284,7 +266,7 @@ public class FirebaseModelDownloader {
             firebaseOptions.getProjectId(), modelName, modelHash);
 
     return incomingModelDetails.continueWithTask(
-        executor,
+        bgExecutor,
         incomingModelDetailTask -> {
           if (incomingModelDetailTask.isSuccessful()) {
             // null means we have the latest model or we failed to connect.
@@ -368,7 +350,7 @@ public class FirebaseModelDownloader {
             return fileDownloadService
                 .download(incomingModelDetailTask.getResult(), conditions)
                 .continueWithTask(
-                    executor,
+                    blockingExecutor,
                     downloadTask -> {
                       if (downloadTask.isSuccessful()) {
                         return finishModelDownload(modelName);
@@ -401,14 +383,14 @@ public class FirebaseModelDownloader {
               firebaseOptions.getProjectId(), modelName);
       // no local model - start download.
       return retryModelDetails.continueWithTask(
-          executor,
+          bgExecutor,
           retryModelDetailTask -> {
             if (retryModelDetailTask.isSuccessful()) {
               // start download
               return fileDownloadService
                   .download(retryModelDetailTask.getResult(), conditions)
                   .continueWithTask(
-                      executor,
+                      bgExecutor,
                       retryDownloadTask -> {
                         if (retryDownloadTask.isSuccessful()) {
                           return finishModelDownload(modelName);
@@ -458,7 +440,7 @@ public class FirebaseModelDownloader {
     fileDownloadService.maybeCheckDownloadingComplete();
 
     TaskCompletionSource<Set<CustomModel>> taskCompletionSource = new TaskCompletionSource<>();
-    executor.execute(
+    bgExecutor.execute(
         () -> taskCompletionSource.setResult(sharedPreferencesUtil.listDownloadedModels()));
     return taskCompletionSource.getTask();
   }
@@ -472,7 +454,7 @@ public class FirebaseModelDownloader {
   public Task<Void> deleteDownloadedModel(@NonNull String modelName) {
 
     TaskCompletionSource<Void> taskCompletionSource = new TaskCompletionSource<>();
-    executor.execute(
+    bgExecutor.execute(
         () -> {
           // remove all files associated with this model and then clean up model references.
           boolean isSuccessful = deleteModelDetails(modelName);